refactor reporting batch worker routes
